Conflicts:
	BDT-JPA/src/cl/minsal/divap/model/Componente.java
	BDT-JPA/src/cl/minsal/divap/model/Plantilla.java
	BDT-JPA/src/cl/minsal/divap/model/ProgramaAno.java
	BDT-JPA/src/cl/minsal/divap/model/ProgramaMunicipalCore.java
	BDT-ejb/ejbModule/minsal/divap/dao/DocumentDAO.java
	BDT-ejb/ejbModule/minsal/divap/dao/ProgramasDAO.java
	BDT-ejb/ejbModule/minsal/divap/excel/GeneradorExcel.java
	BDT-ejb/ejbModule/minsal/divap/service/ProgramasService.java
	BDT-web/src/main/java/cl/minsal/divap/controller/ProcesoEstimacionFlujoCajaRevisarValidarMonitoreoController.java
	script/script_creacion_modelo V8.0.sql

27 de octubre

+++ BDT-JPA/src/cl/minsal/divap/model/CajaMontoPK.java
@@ -49,8 +49,8 @@ public class CajaMontoPK implements Serializable {
     @Override
     public int hashCode() {
         int hash = 0;
-        hash += (int) caja;
-        hash += (int) mes;
+        hash += caja;
+        hash += mes;
         return hash;
     }
